Switch from using GOV.UK Jenkins to Travis
This project isn't owned by GOV.UK Publishing anymore so it needs
to stop using Jenkins to publish changes.

Ideally I'd use a deploy key for this, but it turns out that you can't
attach the same deploy key to more than one repo. This means that we're
going to use a GitHub token attached to the govuk-ci user - this should
be moved to a different user in the future. To make that happen:

- Add the personal access token as a secure Heroku env var
- Change the Git URLs to push over HTTPS, not SSH

The gem_publisher gem expects there to be a file at `.gem/credentials`.
I've encrypted that file for Travis and added that file to the Travis
worker.

Same deal for the `npm publish` command.

It turns out you can only encrypt one file at a time which means
that we need to create a tar archive and then move things around
afterwards which is a little bit horrible.

In order to do this I made a secrets directory and added 2 files:

gem_credentials:

```
---
:rubygems_api_key: redacted
```

npmrc:

```
_auth = redacted
email = name@example.com
```

And then ran:

```
tar cvf secrets.tar secrets/npmrc secrets/gem_credentials
```

And encrypted that file following these docs:
https://docs.travis-ci.com/user/encrypting-files/